GeneralizedChisqDistribution

This package implements the Generalized Chi-squared distribution in Julia.

Installation

Add this package to the current environment with ]add GeneralizedChisqDistribution in the "pkg mode" of the REPL, or in the "standard REPL":

using Pkg
Pkg.add("GeneralizedChisqDistribution")

Usage

GeneralizedChisqDistribution extends the functionality of Distributions.jl by exporting the univariate continuous distribution GeneralizedChisq.

For example, calculate the CDF at x = 5.0 of a distribution that represents the weighted sum of two noncentral Chi-squared and a Normal variable with the following parameters:

Weights (w) of the Chi-squared distributions equal to 1 and -1.

Degrees of freedom (ν) 1 and 2.

Noncentrality parameters (λ) equal to 1.5 for both distibutions.

Mean of the normal variable (μ) equal to 10.

Standard deviation of the normal variable (σ) equal to 1.

using Distributions
using GeneralizedChisqDistribution

d = GeneralizedChisq([1,-1], [1,2], [1.5, 1.5], 10, 1)

cdf(d, 5.0)

API

GeneralizedChisqDistribution.GeneralizedChisq — Type

GeneralizedChisq(w, ν, λ, μ, σ)

The Generalized chi-squared distribution is the distribution of a sum of independent noncentral chi-squared variables and a normal variable:

\[\xi =\sum_{i}w_{i}y_{i}+x,\quad y_{i}\sim \chi^{2}(\nu_{i},\lambda _{i}),\quad x\sim N(\mu,\sigma^{2}).\]

The variables w, ν, and λ are vectors with the weights, degrees of freedom and noncentrality parameters of the Chi-squared distributions, respectively. μ is the mean and σ is the standard deviation of the normal distribution.

External links

Generalized chi-squared distribution on Wikipedia